+++ src/debugFunction.ts
@@ -60,18 +60,9 @@ export function postFamiliarRender(_familiar: EntityFamiliar): void {
 
 // ModCallbacks.MC_POST_NPC_RENDER (28)
 export function postNPCRender(_npc: EntityNPC): void {
-  /*
-  if (npc.Parent !== undefined) {
-    return;
-  }
-
-  const text = `State: ${npc.State}, StateFrame: ${npc.StateFrame}, I1: ${npc.I1}, I2: ${npc.I2}, V1: ${npc.V1}, V2: ${npc.V2}`;
-  const sprite = npc.GetSprite();
-  const animation = sprite.GetAnimation();
-  const text = `Animation: ${animation}`;
-  const position = Isaac.WorldToScreen(npc.Position);
-  Isaac.RenderText(text, position.X, position.Y, 1, 1, 1, 1);
-  */
+  // const text = `State: ${npc.State}, StateFrame: ${npc.StateFrame}, I1: ${npc.I1}, I2: ${npc.I2}, V1: ${npc.V1}, V2: ${npc.V2}`;
+  // const position = Isaac.WorldToScreen(npc.Position);
+  // Isaac.RenderText(text, position.X, position.Y, 1, 1, 1, 1);
 }
 
 // ModCallbacks.MC_POST_PLAYER_RENDER (32)