+++ src/features/optional/enemies/fastPolties.ts
@@ -1,0 +1,11 @@
+import { config } from "../../../modConfigMenu";
+
+// ModCallbacksCustom.MC_POST_NPC_INIT_LATE
+// EntityType.ENTITY_POLTY (816)
+export function postNPCInitLatePolty(npc: EntityNPC): void {
+  if (!config.fastPolties) {
+    return;
+  }
+
+  npc.StateFrame = 0;
+}